# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: TODO API
  version: 1.0.0
  description: A simple production-ready public API

servers:
  - url: http://localhost:4000
    description: Production Server
  - url: http://localhost:4000
    description: Local Development Server

paths:
  /todos:
    get:
      summary: Get all TODOs
      operationId: getTodos
      tags:
        - Todos
      responses:
        '200':
          description: A list of TODO items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new TODO
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '201':
          description: TODO created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "Learn OpenAPI"
        completed:
          type: boolean
          example: false

    Error:
      type: object
      properties:
        error:
          type: string
          example: "ID and title are required"